[🚚rename]: 컴포넌트 Common 폴더로  이동
