+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export { Avatar, EditButton } from './Avatar';
+export { Avatar, EditButton } from './Common/Avatar';
 export { default as Modal } from './Common/Modal/Modal';
 export { default as Spinner } from './Common/Spinner/Spinner';
 export { default as ThemeContext } from './ThemeContext';
